WIP: Raise exception on invalid service
This commit raises exception, if a call is invoked on invalid
service. This should enable the tests disabled as proper exception
is not raised for invalid datasource.

TODO: If the exceptions are raised from endpoints, they are wrapped
into RemoteError and raised, so this needs to be handled
either remove RemoteError and raise original exception or
handle remote exception on API for tests to pass

Change-Id: I1deb4d350e8be90c9c8a76e0b5c7c9e7051405a8