WIP: Migrate harness to DSE2
Pull API models off the bus and give them an argument
to the bus so they can still RPC.  Necessary since API models
are part of 1 service on the bus in DSE2 (and never needed to be
on the bus for DSE1).

Introduce new creation process that bootstraps system using DSE2.

Change-Id: Ic91e2f7918970766fad8e2aff4a0ef88fb5777e2