Launch API Server and/or Server based on command
Enables congress to start with start_process.py and to switch if
it has API service and/or PolicyEngine in the node or doesn't have.
congress.conf keeps datasource driver list the node has. And Node
would have policy engine service and api service based on its
command line arguments.

The ways to launch Congress are following:
In case of distributed_architecture=False (for old arhitcture and same command):
  $ python congress-server --config-file /path/to/congress.conf

In case of distributed_architecture=True:
  - launch with datasource service only:
    $ python congress-server --config-file /path/to/congress.conf --without-engine --without-api
  - launch with API services:
    $ python congress-server --config-file /path/to/congress.conf --without-engine
  - launch with Policy Engine:
    $ python congress-server --config-file /path/to/congress.conf --without-api
  - launch with both API and Policy Engine:
    $ python congress-server --config-file /path/to/congress.conf

Change-Id: Icf13f71bf4748ed3ef2e1821776d22c7862945bd
Partial-Bug: #1541019